ml-api/requirements.txt
# ml-api/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
google-generativeai
geopy
folium
python-multipart
python-dotenv
pydantic
httpx
pymongo